requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
firebase-admin==6.1.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart
sqlalchemy
mysqlclient
mysql-connector-python
pytest
faker